Fixup types.

+++ src/Game.cpp
@@ -22,8 +22,8 @@ bool Game::init(const std::string& _name) {
         LOG(ERROR) << "Can't load '" << iniFilename << "', using defaults";
     } else {
         // 1200x675 is a 16:9 window that fits inside a 1366x768 screen on most systems
-        config.width = (unsigned int) abs(reader.GetInteger("game", "width", config.width));
-        config.height = (unsigned int) abs(reader.GetInteger("game", "height", config.height));
+        config.width = (uint32_t) abs(reader.GetInteger("game", "width", static_cast<long>(config.width)));
+        config.height = (uint32_t) abs(reader.GetInteger("game", "height", static_cast<long>(config.height)));
         config.fullscreen = reader.GetBoolean("game", "fullscreen", config.fullscreen);
         config.useDesktopSize = reader.GetBoolean("game", "useDesktopSize", config.useDesktopSize);
         config.deadZone = reader.GetReal("game", "deadZone", config.deadZone);
